Extract /model switcher logic into agent.core.model_switcher
main.py was accumulating model-switch specifics (suggested list, id
format check, HF routing info printer, probe-and-switch flow, commit
helper). Moving them to a dedicated module keeps the REPL dispatcher
focused on input parsing and makes the switcher independently testable.

Net: main.py down ~160 lines, /model handler is now a 10-line delegation.
No behavior change.
