+++ agent/main.py
@@ -22,6 +22,7 @@ from prompt_toolkit import PromptSession
 
 from agent.config import load_config
 from agent.core.agent_loop import submission_loop
+from agent.core import model_switcher
 from agent.core.session import OpType
 from agent.core.tools import ToolRouter
 from agent.utils.reliability_checks import check_training_script_save_pattern
@@ -49,40 +50,6 @@ litellm.drop_params = True
 # on every error — users don't need it, and our friendly errors cover the case.
 litellm.suppress_debug_info = True
 
-# ── Suggested models shown by `/model` (not a gate) ──────────────────────
-# Users can paste any HF model id (e.g. "MiniMaxAI/MiniMax-M2.7") or use one
-# of the `anthropic/` / `openai/` prefixes for direct API access. For HF ids,
-# append ":fastest" / ":cheapest" / ":preferred" / ":<provider>" to override
-# the default routing policy (auto = fastest with failover).
-SUGGESTED_MODELS = [
-    {"id": "anthropic/claude-opus-4-7", "label": "Claude Opus 4.7"},
-    {"id": "anthropic/claude-opus-4-6", "label": "Claude Opus 4.6"},
-    {"id": "MiniMaxAI/MiniMax-M2.7", "label": "MiniMax M2.7"},
-    {"id": "moonshotai/Kimi-K2.6", "label": "Kimi K2.6"},
-    {"id": "zai-org/GLM-5.1", "label": "GLM 5.1"},
-]
-
-
-def _is_valid_model_id(model_id: str) -> bool:
-    """Loose format check — lets users pick any model id.
-
-    Accepts:
-      • anthropic/<model>
-      • openai/<model>
-      • <org>/<model>[:<tag>]            (HF router; tag = provider or policy)
-      • huggingface/<org>/<model>[:<tag>] (same, accepts legacy prefix)
-
-    Actual availability is verified against the HF router catalog on switch,
-    or by the provider on first call.
-    """
-    if not model_id or "/" not in model_id:
-        return False
-    # Strip :tag suffix before structural check
-    head = model_id.split(":", 1)[0]
-    parts = head.split("/")
-    return len(parts) >= 2 and all(parts)
-
-
 def _safe_get_args(arguments: dict) -> dict:
     """Safely extract args dict from arguments, handling cases where LLM passes string."""
     args = arguments.get("args", {})
@@ -90,159 +57,6 @@ def _safe_get_args(arguments: dict) -> dict:
     if isinstance(args, str):
         return {}
     return args if isinstance(args, dict) else {}
-
-
-_ROUTING_POLICIES = {"fastest", "cheapest", "preferred"}
-
-
-def _print_hf_routing_info(model_id: str, console) -> bool:
-    """Show HF router catalog info (providers, price, context, tool support)
-    for an HF-router model id. Returns ``True`` to signal the caller can
-    proceed with the switch, ``False`` to indicate a hard problem the user
-    should notice before we fire the effort probe.
-
-    Anthropic / OpenAI ids return ``True`` without printing anything
-    (the probe below covers "does this model exist").
-    """
-    if model_id.startswith(("anthropic/", "openai/")):
-        return True
-
-    from agent.core import hf_router_catalog as cat
-
-    bare, _, tag = model_id.partition(":")
-    info = cat.lookup(bare)
-    if info is None:
-        console.print(
-            f"[bold red]Warning:[/bold red] '{bare}' isn't in the HF router "
-            "catalog. Checking anyway — first call may fail."
-        )
-        suggestions = cat.fuzzy_suggest(bare)
-        if suggestions:
-            console.print(f"[dim]Did you mean: {', '.join(suggestions)}[/dim]")
-        return True
-
-    live = info.live_providers
-    if not live:
-        console.print(
-            f"[bold red]Warning:[/bold red] '{bare}' has no live providers "
-            "right now. First call will likely fail."
-        )
-        return True
-
-    if tag and tag not in _ROUTING_POLICIES:
-        matched = [p for p in live if p.provider == tag]
-        if not matched:
-            names = ", ".join(p.provider for p in live)
-            console.print(
-                f"[bold red]Warning:[/bold red] provider '{tag}' doesn't serve "
-                f"'{bare}'. Live providers: {names}. Checking anyway."
-            )
-
-    if not info.any_supports_tools:
-        console.print(
-            f"[bold red]Warning:[/bold red] no provider for '{bare}' advertises "
-            "tool-call support. This agent relies on tool calls — expect errors."
-        )
-
-    if tag in _ROUTING_POLICIES:
-        policy = tag
-    elif tag:
-        policy = f"pinned to {tag}"
-    else:
-        policy = "auto (fastest)"
-    console.print(f"  [dim]routing: {policy}[/dim]")
-    for p in live:
-        price = (
-            f"${p.input_price:g}/${p.output_price:g} per M tok"
-            if p.input_price is not None and p.output_price is not None
-            else "price n/a"
-        )
-        ctx = f"{p.context_length:,} ctx" if p.context_length else "ctx n/a"
-        tools = "tools" if p.supports_tools else "no tools"
-        console.print(
-            f"  [dim]{p.provider}: {price}, {ctx}, {tools}[/dim]"
-        )
-    return True
-
-
-async def _probe_and_switch_model(
-    model_id: str,
-    config,
-    session,
-    console,
-) -> None:
-    """Validate model+effort with a 1-token ping, cache the effective effort,
-    then commit the switch.
-
-    Three visible outcomes:
-
-    * ✓ ``effort: <level>`` — model accepted the preferred effort (or a
-      fallback from the cascade; the note explains if so)
-    * ✓ ``effort: off`` — model doesn't support thinking; we'll strip it
-    * ✗ hard error (auth, model-not-found, quota) — we reject the switch
-      and keep the current model so the user isn't stranded
-
-    Transient errors (5xx, timeout) complete the switch with a yellow
-    warning; the next real call re-surfaces the error if it's persistent.
-    """
-    from agent.core.effort_probe import ProbeInconclusive, probe_effort
-
-    hf_token = _get_hf_token()
-    preference = config.reasoning_effort
-    if not _print_hf_routing_info(model_id, console):
-        return
-
-    if not preference:
-        # Nothing to validate with a ping that we couldn't validate on the
-        # first real call just as cheaply. Skip the probe entirely.
-        _commit_model_switch(model_id, config, session, effective=None, cache=False)
-        console.print(f"[green]Model switched to {model_id}[/green] [dim](effort: off)[/dim]")
-        return
-
-    console.print(f"[dim]checking {model_id} (effort: {preference})...[/dim]")
-    try:
-        outcome = await probe_effort(model_id, preference, hf_token)
-    except ProbeInconclusive as e:
-        _commit_model_switch(model_id, config, session, effective=None, cache=False)
-        console.print(
-            f"[yellow]Model switched to {model_id}[/yellow] "
-            f"[dim](couldn't validate: {e}; will verify on first message)[/dim]"
-        )
-        return
-    except Exception as e:
-        # Hard persistent error — auth, unknown model, quota. Don't switch.
-        console.print(f"[bold red]Switch failed:[/bold red] {e}")
-        console.print(f"[dim]Keeping current model: {config.model_name}[/dim]")
-        return
-
-    _commit_model_switch(
-        model_id, config, session,
-        effective=outcome.effective_effort, cache=True,
-    )
-    effort_label = outcome.effective_effort or "off"
-    suffix = f" — {outcome.note}" if outcome.note else ""
-    console.print(
-        f"[green]Model switched to {model_id}[/green] "
-        f"[dim](effort: {effort_label}{suffix}, {outcome.elapsed_ms}ms)[/dim]"
-    )
-
-
-def _commit_model_switch(model_id, config, session, effective, cache: bool) -> None:
-    """Apply the switch to the session (or bare config if no session yet).
-
-    ``effective`` is the probe's resolved effort; ``cache=True`` stores it
-    in the session's per-model cache so real calls use the resolved level
-    instead of re-probing. ``cache=False`` (inconclusive probe / effort
-    off) leaves the cache untouched — next call falls back to preference.
-    """
-    if session is not None:
-        session.update_model(model_id)
-        if cache:
-            session.model_effective_effort[model_id] = effective
-        else:
-            session.model_effective_effort.pop(model_id, None)
-    else:
-        config.model_name = model_id
 
 
 def _get_hf_token() -> str | None:
@@ -926,31 +740,16 @@ async def _handle_slash_command(
     if command == "/model":
         console = get_console()
         if not arg:
-            current = config.model_name if config else ""
-            console.print("[bold]Current model:[/bold]")
-            console.print(f"  {current}")
-            console.print("\n[bold]Suggested:[/bold]")
-            for m in SUGGESTED_MODELS:
-                marker = " [dim]<-- current[/dim]" if m["id"] == current else ""
-                console.print(f"  {m['id']}  [dim]({m['label']})[/dim]{marker}")
-            console.print(
-                "\n[dim]Paste any HF model id (e.g. 'MiniMaxAI/MiniMax-M2.7').\n"
-                "Add ':fastest', ':cheapest', ':preferred', or ':<provider>' to override routing.\n"
-                "Use 'anthropic/<model>' or 'openai/<model>' for direct API access.[/dim]"
-            )
+            model_switcher.print_model_listing(config, console)
             return None
-        if not _is_valid_model_id(arg):
-            console.print(f"[bold red]Invalid model id format:[/bold red] {arg}")
-            console.print(
-                "[dim]Expected:\n"
-                "  • <org>/<model>[:tag]    (HF router — paste from huggingface.co)\n"
-                "  • anthropic/<model>\n"
-                "  • openai/<model>[/dim]"
-            )
+        if not model_switcher.is_valid_model_id(arg):
+            model_switcher.print_invalid_id(arg, console)
             return None
         normalized = arg.removeprefix("huggingface/")
         session = session_holder[0] if session_holder else None
-        await _probe_and_switch_model(normalized, config, session, console)
+        await model_switcher.probe_and_switch_model(
+            normalized, config, session, console, _get_hf_token(),
+        )
         return None
 
     if command == "/yolo":
